fix [screensaver]: prevent unintended screensaver activation by aborting in 'unsafe' state
This commit fixes a problem where the vimatrix screensaver autocommands
would be blocked from resetting the screensaver timer, due to vim having
entered an 'unsafe' state, thus scheduling the timer unintentionally,
e.g. in an unfocused neovim session. The fix simply checks for a 'safe'
state upon timer expiry and aborts when state is reported as not 'safe'.

It might add a side-effect of missing a screensaver execution if an
'unsafe' state is not triggered by the user, I am not aware of any
examples of this happening and here choose for a simple, effective, and
safe solution to the observed problem.